admin remove flight API

+++ backend/app/api/resources/admin_delete_api.py
@@ -35,6 +35,14 @@ class AdminDeleteModel(Resource):
                 if not validate_model_id(int(model_id)):
                     raise Exception("Invalid search parameters")
                 
+                # check if model is in the database
+                query = "SELECT * FROM model WHERE Model_ID = %s"
+                cursor.execute(query,(int(model_id),))
+                items = cursor.fetchone()
+
+                if items is None:
+                    raise Exception("404")
+                
                 query = """
                     DELETE FROM model WHERE Model_ID = %s
                 """
@@ -48,6 +56,8 @@ class AdminDeleteModel(Resource):
             except Exception as ex:
                 if str(ex) == "403":
                     return abort(403, message="Only admins can access API")
+                if str(ex) == "404":
+                    return abort(404, message="Model not found")
                 return abort(400, message=f"Failed to Delete Model. Error: {ex}")
         else:
             return abort(403, message="Unauthorized Access")
@@ -83,6 +93,14 @@ class AdminDeleteRoute(Resource):
                 if not validate_route_id(int(route_id)):
                     raise Exception("Invalid search parameters")
                 
+                # check if route is in the database
+                query = "SELECT * FROM route WHERE Route_ID = %s"
+                cursor.execute(query,(int(route_id),))
+                items = cursor.fetchone()
+
+                if items is None:
+                    raise Exception("404")
+                
                 query = """
                     DELETE FROM route WHERE Route_ID = %s
                 """
@@ -96,6 +114,8 @@ class AdminDeleteRoute(Resource):
             except Exception as ex:
                 if str(ex) == "403":
                     return abort(403, message="Only admins can access API")
+                if str(ex) == "404":
+                    return abort(404, message="Route not found")
                 return abort(400, message=f"Failed to Delete Route. Error: {ex}")
         else:
             return abort(403, message="Unauthorized Access")
@@ -131,6 +151,14 @@ class AdminDeleteAirport(Resource):
                 if not validate_icao_code(icao_code):
                     raise Exception("Invalid search parameters")
                 
+                # check if airport is in the database
+                query = "SELECT * FROM airport WHERE ICAO_Code = %s"
+                cursor.execute(query,(icao_code,))
+                items = cursor.fetchone()
+
+                if items is None:
+                    raise Exception("404")
+                
                 query = """
                     DELETE FROM airport WHERE ICAO_Code = %s
                 """
@@ -144,6 +172,8 @@ class AdminDeleteAirport(Resource):
             except Exception as ex:
                 if str(ex) == "403":
                     return abort(403, message="Only admins can access API")
+                if str(ex) == "404":
+                    return abort(404, message="Airport not found")
                 return abort(400, message=f"Failed to Delete Airport. Error: {ex}")
         else:
             return abort(403, message="Unauthorized Access")
@@ -179,6 +209,14 @@ class AdminDeleteAirplane(Resource):
                 if not validate_tail_number(tail_number):
                     raise Exception("Invalid search parameters")
                 
+                # check if airplane is in the database
+                query = "SELECT * FROM airplane WHERE Tail_Number = %s"
+                cursor.execute(query,(tail_number,))
+                items = cursor.fetchone()
+
+                if items is None:
+                    raise Exception("404")
+                
                 query = """
                     DELETE FROM airplane WHERE Tail_Number = %s
                 """
@@ -192,6 +230,8 @@ class AdminDeleteAirplane(Resource):
             except Exception as ex:
                 if str(ex) == "403":
                     return abort(403, message="Only admins can access API")
+                if str(ex) == "404":
+                    return abort(404, message="Airplane not found")
                 return abort(400, message=f"Failed to Delete Airplane. Error: {ex}")
         else:
             return abort(403, message="Unauthorized Access")
@@ -227,13 +267,20 @@ class AdminDeleteScheduledFlight(Resource):
                 if not validate_flight_id(int(flight_id)):
                     raise Exception("Invalid search parameters")
                 
+                # check if scheduled flight is in the database
+                query = "SELECT * FROM scheduled_flight WHERE Scheduled_ID = %s"
+                cursor.execute(query,(int(flight_id),))
+                items = cursor.fetchone()
+
+                if items is None:
+                    raise Exception("404")
+                
                 query = """
                     DELETE FROM scheduled_flight WHERE Scheduled_ID = %s
                 """
 
                 # Execute query with scheduled flight id
                 cursor.execute(query,(int(flight_id),))
-                print("reached")
                 connection.commit()
                 connection.close()
                 
@@ -241,17 +288,66 @@ class AdminDeleteScheduledFlight(Resource):
             except Exception as ex:
                 if str(ex) == "403":
                     return abort(403, message="Only admins can access API")
-                if (ex.args[1]=='1644 (45000): Flight has paid bookings'):
-                    try:
-                        cursor.execute("""
-                            UPDATE scheduled_flight SET Departure_Time = '2000-01-01 00:00:00' WHERE Scheduled_ID = %s;
-                        """, (int(flight_id),))
-                        connection.commit()
-                        connection.close()
-                        return make_response("Scheduled Flight was not deleted. Hidden From Users", 409)
-                    except Exception as ex:
-                        raise Exception(f"Failed to reschedule")
+                if str(ex) == "404":
+                    return abort(404, message="Scheduled Flight not found")
                 return abort(400, message=f"Failed to Delete Scheduled Flight. Error: {ex}")
         else:
             return abort(403, message="Unauthorized Access")
 
+
+class AdminHideScheduledFlight(Resource):
+    @jwt_required()
+    def patch(self, flight_id):
+        try:
+            connection = get_db_connection_admin()
+        except Exception as ex:
+            return abort(500, message=f"Failed to connect to database. Error: {ex}")
+        
+        if connection:
+            try:
+                cursor = connection.cursor(prepared=True)
+
+                # Get current user
+                current_user = get_jwt_identity()
+
+                query = """
+                    SELECT * FROM staff WHERE Username = %s AND Role = 'Admin'
+                """
+
+                # Execute query with username
+                cursor.execute(query,(current_user,))
+                items = cursor.fetchone()
+
+                if items is None:
+                    raise Exception("403")
+                
+                # validate search parameters
+                if not validate_flight_id(int(flight_id)):
+                    raise Exception("Invalid search parameters")
+                
+                # check if scheduled flight is in the database
+                query = "SELECT * FROM scheduled_flight WHERE Scheduled_ID = %s"
+                cursor.execute(query,(int(flight_id),))
+                items = cursor.fetchone()
+
+                if items is None:
+                    raise Exception("404")
+                
+                query = """
+                    UPDATE scheduled_flight SET Departure_Time = '2000-01-01 00:00:00' WHERE Scheduled_ID = %s;
+                """
+
+                # Execute query with scheduled flight id
+                cursor.execute(query,(int(flight_id),))
+                connection.commit()
+                connection.close()
+                
+                return make_response("Scheduled Flight Hidden Successfully", 200)
+            except Exception as ex:
+                if str(ex) == "403":
+                    return abort(403, message="Only admins can access API")
+                if str(ex) == "404":
+                    return abort(404, message="Scheduled Flight not found")
+                return abort(400, message=f"Failed to Hide Scheduled Flight. Error: {ex}")
+        else:
+            return abort(403, message="Unauthorized Access")
